Task2: get the bit of a number at a given index, fix for: index out of range

+++ 0x14-bit_manipulation/main.h
@@ -3,11 +3,12 @@
 
 /* Libraries */
 #include <stdio.h>
+#include <limits.h>
 
 /* Funtion prototypes */
 int _putchar(char c);
 unsigned int binary_to_uint(const char *b);
 void print_binary(unsigned long int n);
-int get_bit(unsigned int n, unsigned int index);
+int get_bit(unsigned long int n, unsigned int index);
 
 #endif /* MAIN_H */
